+++ src/idc/writer/imgcls/_adams.py
@@ -4,12 +4,12 @@ from typing import List
 
 from wai.logging import LOGGING_WARNING
 from wai.common.file.report import Report, Field, save
-from idc.api import ImageClassificationData, SplittableStreamWriter, make_list
+from idc.api import ImageClassificationData, SplittableStreamWriter, make_list, AnnotationsOnlyWriter, add_annotations_only_param
 
 
-class AdamsImageClassificationWriter(SplittableStreamWriter):
+class AdamsImageClassificationWriter(SplittableStreamWriter, AnnotationsOnlyWriter):
 
-    def __init__(self, output_dir: str = None, class_field: str = None,
+    def __init__(self, output_dir: str = None, class_field: str = None, annotations_only: bool = None,
                  split_names: List[str] = None, split_ratios: List[int] = None,
                  logger_name: str = None, logging_level: str = LOGGING_WARNING):
         """
@@ -31,6 +31,7 @@ class AdamsImageClassificationWriter(SplittableStreamWriter):
         super().__init__(split_names=split_names, split_ratios=split_ratios, logger_name=logger_name, logging_level=logging_level)
         self.output_dir = output_dir
         self.class_field = class_field
+        self.annotations_only = annotations_only
 
     def name(self) -> str:
         """
@@ -60,6 +61,7 @@ class AdamsImageClassificationWriter(SplittableStreamWriter):
         parser = super()._create_argparser()
         parser.add_argument("-o", "--output", type=str, help="The directory to store the images/.report files in. Any defined splits get added beneath there.", required=True)
         parser.add_argument("-c", "--class_field", metavar="FIELD", type=str, default=None, help="The report field containing the image classification label", required=True)
+        add_annotations_only_param(parser)
         return parser
 
     def _apply_args(self, ns: argparse.Namespace):
@@ -72,6 +74,7 @@ class AdamsImageClassificationWriter(SplittableStreamWriter):
         super()._apply_args(ns)
         self.output_dir = ns.output
         self.class_field = ns.class_field
+        self.annotations_only = ns.annotations_only
 
     def accepts(self) -> List:
         """
@@ -90,6 +93,8 @@ class AdamsImageClassificationWriter(SplittableStreamWriter):
         if not os.path.exists(self.output_dir):
             self.logger().info("Creating output dir: %s" % self.output_dir)
             os.makedirs(self.output_dir)
+        if self.annotations_only is None:
+            self.annotations_only = False
 
     def write_stream(self, data):
         """
@@ -117,8 +122,9 @@ class AdamsImageClassificationWriter(SplittableStreamWriter):
                     report.set_value(Field.parse_field(k), item.get_metadata()[k])
 
             path = os.path.join(sub_dir, item.image_name)
-            self.logger().info("Writing image to: %s" % path)
-            item.save_image(path)
+            if not self.annotations_only:
+                self.logger().info("Writing image to: %s" % path)
+                item.save_image(path)
 
             if not empty:
                 path = os.path.splitext(path)[0] + ".report"