+++ src/idc/writer/objdet/_opex.py
@@ -4,13 +4,13 @@ from datetime import datetime
 from typing import List
 
 from wai.logging import LOGGING_WARNING
-from idc.api import ObjectDetectionData, SplittableStreamWriter, get_object_label, make_list
+from idc.api import ObjectDetectionData, SplittableStreamWriter, get_object_label, make_list, AnnotationsOnlyWriter, add_annotations_only_param
 from opex import ObjectPredictions, ObjectPrediction, BBox, Polygon
 
 
-class OPEXObjectDetectionWriter(SplittableStreamWriter):
+class OPEXObjectDetectionWriter(SplittableStreamWriter, AnnotationsOnlyWriter):
 
-    def __init__(self, output_dir: str = None,
+    def __init__(self, output_dir: str = None, annotations_only: bool = None,
                  split_names: List[str] = None, split_ratios: List[int] = None,
                  logger_name: str = None, logging_level: str = LOGGING_WARNING):
         """
@@ -18,6 +18,8 @@ class OPEXObjectDetectionWriter(SplittableStreamWriter):
 
         :param output_dir: the output directory to save the image/report in
         :type output_dir: str
+        :param annotations_only: whether to output only the annotations and not the images
+        :type annotations_only: bool
         :param split_names: the names of the splits, no splitting if None
         :type split_names: list
         :param split_ratios: the integer ratios of the splits (must sum up to 100)
@@ -29,6 +31,7 @@ class OPEXObjectDetectionWriter(SplittableStreamWriter):
         """
         super().__init__(split_names=split_names, split_ratios=split_ratios, logger_name=logger_name, logging_level=logging_level)
         self.output_dir = output_dir
+        self.annotations_only = annotations_only
 
     def name(self) -> str:
         """
@@ -57,6 +60,7 @@ class OPEXObjectDetectionWriter(SplittableStreamWriter):
         """
         parser = super()._create_argparser()
         parser.add_argument("-o", "--output", type=str, help="The directory to store the images/.json files in. Any defined splits get added beneath there.", required=True)
+        add_annotations_only_param(parser)
         return parser
 
     def _apply_args(self, ns: argparse.Namespace):
@@ -68,6 +72,7 @@ class OPEXObjectDetectionWriter(SplittableStreamWriter):
         """
         super()._apply_args(ns)
         self.output_dir = ns.output
+        self.annotations_only = ns.annotations_only
 
     def accepts(self) -> List:
         """
@@ -86,6 +91,8 @@ class OPEXObjectDetectionWriter(SplittableStreamWriter):
         if not os.path.exists(self.output_dir):
             self.logger().info("Creating output dir: %s" % self.output_dir)
             os.makedirs(self.output_dir)
+        if self.annotations_only is None:
+            self.annotations_only = False
 
     def write_stream(self, data):
         """
@@ -110,8 +117,9 @@ class OPEXObjectDetectionWriter(SplittableStreamWriter):
             path = sub_dir
             os.makedirs(path, exist_ok=True)
             path = os.path.join(path, item.image_name)
-            self.logger().info("Writing image to: %s" % path)
-            item.save_image(path)
+            if not self.annotations_only:
+                self.logger().info("Writing image to: %s" % path)
+                item.save_image(path)
 
             # annotations
             if absolute is not None: