+++ src/idc/writer/objdet/_voc.py
@@ -4,7 +4,7 @@ from typing import List
 from xml.etree.ElementTree import Element, ElementTree
 
 from wai.logging import LOGGING_WARNING
-from idc.api import ObjectDetectionData, SplittableStreamWriter, get_object_label, make_list
+from idc.api import ObjectDetectionData, SplittableStreamWriter, get_object_label, make_list, AnnotationsOnlyWriter, add_annotations_only_param
 
 
 def append_element(to: Element, tag: str, value):
@@ -20,9 +20,9 @@ def append_element(to: Element, tag: str, value):
     to.append(element)
 
 
-class VOCObjectDetectionWriter(SplittableStreamWriter):
+class VOCObjectDetectionWriter(SplittableStreamWriter, AnnotationsOnlyWriter):
 
-    def __init__(self, output_dir: str = None,
+    def __init__(self, output_dir: str = None, annotations_only: bool = None,
                  split_names: List[str] = None, split_ratios: List[int] = None,
                  logger_name: str = None, logging_level: str = LOGGING_WARNING):
         """
@@ -30,6 +30,8 @@ class VOCObjectDetectionWriter(SplittableStreamWriter):
 
         :param output_dir: the output directory to save the image/report in
         :type output_dir: str
+        :param annotations_only: whether to output only the annotations and not the images
+        :type annotations_only: bool
         :param split_names: the names of the splits, no splitting if None
         :type split_names: list
         :param split_ratios: the integer ratios of the splits (must sum up to 100)
@@ -41,6 +43,7 @@ class VOCObjectDetectionWriter(SplittableStreamWriter):
         """
         super().__init__(split_names=split_names, split_ratios=split_ratios, logger_name=logger_name, logging_level=logging_level)
         self.output_dir = output_dir
+        self.annotations_only = annotations_only
 
     def name(self) -> str:
         """
@@ -69,6 +72,7 @@ class VOCObjectDetectionWriter(SplittableStreamWriter):
         """
         parser = super()._create_argparser()
         parser.add_argument("-o", "--output", type=str, help="The directory to store the images/.xml files in. Any defined splits get added beneath there.", required=True)
+        add_annotations_only_param(parser)
         return parser
 
     def _apply_args(self, ns: argparse.Namespace):
@@ -80,6 +84,7 @@ class VOCObjectDetectionWriter(SplittableStreamWriter):
         """
         super()._apply_args(ns)
         self.output_dir = ns.output
+        self.annotations_only = ns.annotations_only
 
     def accepts(self) -> List:
         """
@@ -98,6 +103,8 @@ class VOCObjectDetectionWriter(SplittableStreamWriter):
         if not os.path.exists(self.output_dir):
             self.logger().info("Creating output dir: %s" % self.output_dir)
             os.makedirs(self.output_dir)
+        if self.annotations_only is None:
+            self.annotations_only = False
 
     def _append_element(self, to: Element, tag: str, value):
         """
@@ -134,8 +141,9 @@ class VOCObjectDetectionWriter(SplittableStreamWriter):
             path = sub_dir
             os.makedirs(path, exist_ok=True)
             path = os.path.join(path, item.image_name)
-            self.logger().info("Writing image to: %s" % path)
-            item.save_image(path)
+            if not self.annotations_only:
+                self.logger().info("Writing image to: %s" % path)
+                item.save_image(path)
 
             # annotations
             if absolute is not None: